ai-health-ml/requirements.txt
numpy
pandas
scikit-learn
joblib
fastapi
uvicorn
pydantic
httpx
requests
beautifulsoup4
lxml
schedule
python-dotenv
